Fix for autogeneration of gen-init files when saving init.org
The problem was when comparing the buffer name with the expected file
name. One was using the absolute path while the other was using the
absoulte path using tilde (~). Now we will only check if the file name
is "init.org". Hopefully no other file will be called like that